+++ packages/terra-form-textarea/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Changed
 * Updated "test:jest" script to work on windows.
 
+### Added
+* Ability to get textarea ref via `setRef` prop
+
 2.12.0 - (May 30, 2018)
 ------------------
 ### Changed